+++ src/ui/App.js
@@ -1,10 +1,10 @@
 import styled from 'styled-components'
 import React, { useState } from 'react';
-import generate from './generation-logic/generate'
-import { Media } from 'react-breakpoints';
+import generate from '../generation-logic/generate'
 import breakpoints from './breakpoints';
 import ShareButtons from './ShareButtons';
-
+import AppHeader from './AppHeader'
+import { shuffleArray } from '../utils'
 
 const Container = styled.div`
   height: 100vh;
@@ -15,120 +15,6 @@ const Container = styled.div`
   overflow-x: hidden;
   overflow-y: scroll;
   -webkit-overflow-scrolling: touch;
-`
-
-const TopBar = styled.div`
-  background-color: rgb(255, 198, 5);
-  width: 100vw;
-  min-height: 30px;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  @media (min-width: ${breakpoints.md}px) {
-    min-height: 30px;
-  }
-  @media (min-width: ${breakpoints.lg}px) {
-    min-height: 30px;
-  }
-  @media (min-width: ${breakpoints.xxl}px) {
-    min-height: 40px;
-  }
-  @media (min-width: ${breakpoints.xxxl}px) {
-    min-height: 60px;
-  }
-`
-
-const TopBarText = styled.div`
-  font-weight: bold;
-  font-size: 14px;
-  width: 95vw;
-  @media (min-width: ${breakpoints.md}px) {
-    width: 600px;
-  }
-  @media (min-width: ${breakpoints.lg}px) {
-    width: 750px;
-  }
-  @media (min-width: ${breakpoints.xxl}px) {
-    width: 900px;
-    font-size: 22px;
-  }
-  @media (min-width: ${breakpoints.xxxl}px) {
-    width: 1400px;
-    font-size: 32px;
-  }
-`
-
-const LogoContainer = styled.div`
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  // justify-content: center;
-  align-items: center;
-  height: 150px;
-  @media (min-width: ${breakpoints.md}px) {
-    height: 200px;
-  }
-  @media (min-width: ${breakpoints.xl}px) {
-    height: 230px;
-  }
-  @media (min-width: ${breakpoints.xxl}px) {
-    height: 300px;
-  }
-  @media (min-width: ${breakpoints.xxxl}px) {
-    height: 400px;
-  }
-`
-
-const Logo = styled.img`
-  height: 100%;
-`
-
-const Header = styled.div`
-  position: absolute;
-  color: white;
-  font-weight: bold;
-  max-width: 100vw;
-  width: 100vw;
-  text-align: center;
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-  justify-content: center;
-  align-items: center;
-`
-
-const HeaderTitle = styled.div`
-  margin: 6px;
-  font-size: 24px;
-  max-width: 300px;
-  @media (min-width: ${breakpoints.md}px) {
-    max-width: 90vw;
-    font-size: 36px;
-  }
-  @media (min-width: ${breakpoints.xl}px) {
-    font-size: 42px;
-  }
-  @media (min-width: ${breakpoints.xxl}px) {
-    font-size: 56px;
-  }
-  @media (min-width: ${breakpoints.xxxl}px) {
-    font-size: 72px;
-  }
-`
-
-const HeaderSub = styled.p`
-  margin: 6px;
-  font-size: 16px;
-  @media (min-width: ${breakpoints.md}px) {
-    font-size: 20px;
-  }
-  @media (min-width: ${breakpoints.xxl}px) {
-    font-size: 28px;
-  }
-  @media (min-width: ${breakpoints.xxxl}px) {
-    font-size: 40px;
-  }
 `
 
 const SentencesContainer = styled.div`
@@ -289,26 +175,6 @@ const StarIcon = styled.span`
   background-repeat: no-repeat;
 `
 
-
-function shuffle(array) {
-  var currentIndex = array.length, temporaryValue, randomIndex;
-
-  // While there remain elements to shuffle...
-  while (0 !== currentIndex) {
-
-    // Pick a remaining element...
-    randomIndex = Math.floor(Math.random() * currentIndex);
-    currentIndex -= 1;
-
-    // And swap it with the current element.
-    temporaryValue = array[currentIndex];
-    array[currentIndex] = array[randomIndex];
-    array[randomIndex] = temporaryValue;
-  }
-
-  return array;
-}
-
 function App(props) {
   /*
    * Format:
@@ -349,43 +215,27 @@ function App(props) {
 
   const date = (new Date()).toLocaleDateString('pl-PL', dateOpts);
 
-  const shuffledOptionalSentences = shuffle(optionalSentences);
+  shuffleArray(optionalSentences);
   const finalSentences = [
     ...obligatorySentences,
-    shuffledOptionalSentences[0],
-    shuffledOptionalSentences[1]
+    optionalSentences[0],
+    optionalSentences[1]
   ];
+
+  shuffleArray(finalSentences)
 
   const { breakpoints, currentBreakpoint } = props;
 
   return (
       <Container>
-        <TopBar id="topBar">
-          <TopBarText>Koronawirus: ważne informacje</TopBarText>
-        </TopBar>
-        <LogoContainer>
-          <Logo src="logo.jpg" />
-          <Header>
-            <HeaderTitle>Generator obostrzeń COVID-19</HeaderTitle>
-            <Media>
-              {({ breakpoints, currentBreakpoint }) => {
-                console.log(breakpoints[currentBreakpoint])
-                return breakpoints[currentBreakpoint] >= breakpoints.sm ? (
-                  <HeaderSub>Sprawdź, co dzisiaj wolno, a czego nie</HeaderSub>
-                ) : (
-                  <></>
-                )
-              }}
-            </Media>
-          </Header>
-        </LogoContainer>
+        <AppHeader />
         <SentencesContainer>
           <SentenceHeader>
             Kancelaria Prezesa Rady Ministrów informuje, że od {date} roku:
           </SentenceHeader>
           {sentences && (
             <SentenceList>
-            {shuffle(finalSentences).map(it => it(sentences))}
+            {finalSentences.map(it => it(sentences))}
             </SentenceList>
           )}
           <StyledButton onClick={buttonAction}>{buttonText}</StyledButton>
