+++ src/utils.js
@@ -1,0 +1,7 @@
+// Durstenfeld Shuffle from https://stackoverflow.com/a/12646864/6417161
+export function shuffleArray(array) {
+    for (let i = array.length - 1; i > 0; i--) {
+        const j = Math.floor(Math.random() * (i + 1));
+        [array[i], array[j]] = [array[j], array[i]];
+    }
+}
